backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pdfplumber==0.11.0
python-docx==1.1.2
scikit-learn==1.5.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
groq==0.9.0
nltk==3.8.1
pandas==2.2.2
psycopg2-binary==2.9.9
slowapi==0.1.9
prometheus-fastapi-instrumentator==6.1.0
email-validator==2.1.1
